Pipeline_Ativos/requirements.txt
pandas
sqlalchemy
python-dotenv
psycopg2-binary
yfinance
beautifulsoup4==4.13.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
curl_cffi==0.13.0
frozendict==2.4.6
idna==3.10
multitasking==0.0.12
numpy==2.3.2
pandas==2.3.1
peewee==3.18.2
platformdirs==4.3.8
protobuf==6.31.1
psycopg2-binary==2.9.10
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.43
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
yfinance==0.2.65
